refactor: changing file names to streamline w/ code base
