+++ src/practices/Java/JavaMockingFrameworkUsedPractice.ts
