+++ src/practices/Java/JavaPackageManagementUsedPractice.ts
